chore: move pebble specific code back to its own file
