+++ internal/charm/hooks.go
@@ -1,7 +1,6 @@
 package charm
 
 import (
-	"bytes"
 	"crypto/rand"
 	"crypto/tls"
 	"crypto/x509"
@@ -16,7 +15,6 @@ import (
 	"github.com/gruyaume/charm-libraries/prometheus"
 	"github.com/gruyaume/goops"
 	"github.com/gruyaume/notary-k8s-operator/internal/notary"
-	"gopkg.in/yaml.v3"
 )
 
 const (
@@ -605,79 +603,4 @@ func getHostname() string {
 	unitHostname := fmt.Sprintf("%s-%s.%s-endpoints.%s.svc.cluster.local:%d", appName, unitNumber, appName, env.ModelName, APIPort)
 
 	return unitHostname
-}
-
-func pushFile(pebbleClient *client.Client, content string, path string) error {
-	_, err := pebbleClient.SysInfo()
-	if err != nil {
-		return fmt.Errorf("could not connect to pebble: %w", err)
-	}
-
-	source := strings.NewReader(content)
-
-	err = pebbleClient.Push(&client.PushOptions{
-		Source: source,
-		Path:   path,
-	})
-	if err != nil {
-		return fmt.Errorf("could not push file: %w", err)
-	}
-
-	return nil
-}
-
-func getFileContent(pebbleClient *client.Client, path string) (string, error) {
-	target := &bytes.Buffer{}
-
-	err := pebbleClient.Pull(&client.PullOptions{
-		Path:   path,
-		Target: target,
-	})
-	if err != nil {
-		return "", fmt.Errorf("could not get file content: %w", err)
-	}
-
-	return target.String(), nil
-}
-
-type ServiceConfig struct {
-	Override string `yaml:"override"`
-	Summary  string `yaml:"summary"`
-	Command  string `yaml:"command"`
-	Startup  string `yaml:"startup"`
-}
-
-type PebbleLayer struct {
-	Summary     string                   `yaml:"summary"`
-	Description string                   `yaml:"description"`
-	Services    map[string]ServiceConfig `yaml:"services"`
-}
-
-func addPebbleLayer(pebbleClient *client.Client) error {
-	layerData, err := yaml.Marshal(PebbleLayer{
-		Summary:     "Notary layer",
-		Description: "pebble config layer for Notary",
-		Services: map[string]ServiceConfig{
-			"notary": {
-				Override: "replace",
-				Summary:  "Notary Service",
-				Command:  "notary --config " + ConfigPath,
-				Startup:  "enabled",
-			},
-		},
-	})
-	if err != nil {
-		return fmt.Errorf("could not marshal layer data to YAML: %w", err)
-	}
-
-	err = pebbleClient.AddLayer(&client.AddLayerOptions{
-		Combine:   true,
-		Label:     "notary",
-		LayerData: layerData,
-	})
-	if err != nil {
-		return fmt.Errorf("could not add pebble layer: %w", err)
-	}
-
-	return nil
 }
